requirements.txt
click==7.1.2
pyyaml==5.4.1
requests==2.25.1
docker==4.4.4
